refactor : InputType을 CardInputProps의 key로 통일
